+++ components/21-atoms/modal-close/modal-close.config.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  title: 'Modal close',
-  hidden: 'hidden',
-  status: 'ready'
-};