+++ components/31-molecules/fields/checkbox-with-filter/checkbox_filter.functions.js
@@ -128,12 +128,6 @@
     );
 
     /**
-     * Store the button that triggered the modal.
-     * @type {null|Element}
-     */
-    let trigger = null;
-
-    /**
      * Store the checked checkboxes prior to making changes.
      * @type {Array}
      */
@@ -295,51 +289,14 @@
       // hide
       if (modal.classList.contains('visible')) {
         updateCount();
-        openBtn.setAttribute('aria-expanded', 'false');
-        modal.setAttribute('aria-hidden', 'true');
-        modal.classList.remove('visible');
-        if (trigger) {
-          trigger.focus();
-        }
-
         filter(true);
         document.removeEventListener('keydown', handleKeyboardInput);
 
-        /*
-        This component can be part of a filter-organism,
-        we need to remove the scroll lock from the filter modal if visible.
-         */
-        let elem = modal;
-        // eslint-disable-next-line no-empty
-        while ((elem = elem.parentElement) && !elem.classList.contains('modal')) {}
-
-        if (elem && elem.classList.contains('visible')) {
-          elem.style.overflow = '';
-        }
-        else {
-          document.body.style.overflow = '';
-        }
+        history.back();
       }
       // show
       else {
-        openBtn.setAttribute('aria-expanded', 'true');
-        modal.removeAttribute('aria-hidden');
-        document.body.style.overflow = 'hidden';
         document.addEventListener('keydown', handleKeyboardInput);
-        modal.classList.add('visible');
-        modal.focus();
-
-        /*
-        This component can be part of a filter-organism,
-        we need to add a scroll lock to the filter modal if visible.
-         */
-        let elem = modal;
-        // eslint-disable-next-line no-empty
-        while ((elem = elem.parentElement) && !elem.classList.contains('modal')) {}
-
-        if (elem && elem.classList.contains('visible')) {
-          elem.style.overflow = 'hidden';
-        }
       }
     };
 
@@ -382,9 +339,6 @@
      */
     const init = () => {
       selectedFilters = [];
-      modal.setAttribute('tabindex', '-1');
-      modal.setAttribute('aria-hidden', 'true');
-      openBtn.setAttribute('aria-expanded', 'false');
 
       checkboxLoop(({checkbox, label}) => {
         if (checkbox.checked && makeTags) {
@@ -431,9 +385,9 @@
 
       // Enable opening the modal.
       if (openBtn) {
-        openBtn.addEventListener('click', () => {
-          trigger = openBtn;
+        openBtn.addEventListener('click', (e) => {
           selectedFilters = [];
+          window.location.hash = e.target.dataset.hash; // change the hash to open the modal
 
           checkboxLoop(({checkbox}) => {
             if (checkbox.checked) {