+++ components/31-molecules/modal/modal.config.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  title: 'Modal',
+  name: 'modal',
+  status: 'alpha',
+  hidden: true
+};